Fix auth:samctum for API of admin

+++ backend/app/Http/Controllers/VotationController.php
@@ -10,12 +10,20 @@ class VotationController extends Controller
     // READ
     public function index()
     {
+        if (Auth::user()->roleId !== 1) {
+            return response()->json(['error' => 'No tienes permiso'], 403);
+        }
+        
         return response()->json(Votation::all(), 200, [], JSON_UNESCAPED_UNICODE);
     }
 
     // CREATE
     public function store(Request $request)
     {
+        if (Auth::user()->roleId !== 1) {
+            return response()->json(['error' => 'No tienes permiso'], 403);
+        }
+
         $data = $request->validate([
             'title' => 'required|string|max:100',
             'description' => 'nullable|string',
@@ -26,19 +34,19 @@ class VotationController extends Controller
 
         $fakeHash = hash('sha256', uniqid());
 
-    // INSERTO bloque de prueba en la tabla 'block' para que la base de datos no se queje de la Foreign Key
-    //En el siguiente sprint, la blockchain genera los bloques y se guardan en block en la BBDD donde se referencia aquí
-    \DB::table('block')->insertOrIgnore([
-        'hash' => $fakeHash,
-        'blockNumber' => rand(1, 1000),
-        'previousHash' => '0',
-        'transactions' => 0,
-        'isValid' => 1,
-        'createdAt' => now()
-    ]);
+        // INSERTO bloque de prueba en la tabla 'block' para que la base de datos no se queje de la Foreign Key
+        //En el siguiente sprint, la blockchain genera los bloques y se guardan en block en la BBDD donde se referencia aquí
+        \DB::table('block')->insertOrIgnore([
+            'hash' => $fakeHash,
+            'blockNumber' => rand(1, 1000),
+            'previousHash' => '0',
+            'transactions' => 0,
+            'isValid' => 1,
+            'createdAt' => now()
+        ]);
 
-    $data['startBlockHash'] = $fakeHash;
-    $data['endBlockHash'] = $fakeHash;
+        $data['startBlockHash'] = $fakeHash;
+        $data['endBlockHash'] = $fakeHash;
 
         $votation = Votation::create($data);
 
@@ -48,6 +56,10 @@ class VotationController extends Controller
     // UPDATE
     public function update(Request $request, $id)
     {
+        if (Auth::user()->roleId !== 1) {
+            return response()->json(['error' => 'No tienes permiso'], 403);
+        }
+
         $votation = Votation::findOrFail($id);
 
         $votation->update($request->all());
@@ -58,6 +70,10 @@ class VotationController extends Controller
     // DELETE
     public function destroy($id)
     {
+        if (Auth::user()->roleId !== 1) {
+            return response()->json(['error' => 'No tienes permiso'], 403);
+        }
+
         Votation::destroy($id);
         return response()->json(null, 204);
     }
